add nav back to swiping page
